refactor(tests): Organize threat intelligence tests into dedicated directory
- Move provider tests to tests/threat_intelligence/ directory
- Rename test files to follow consistent *_tests.rs naming convention
- Add mod.rs for threat intelligence test module
- Improve test organization and discoverability

Tests:
- All 430 tests passing
- 10 integration tests ignored (requires API keys)
- Test hierarchy matches other test categories (unit, integration, database)
