+++ tests/session_management_tests.rs
