+++ tests/threat_intelligence/mod.rs
@@ -1,0 +1,9 @@
+//! 脅威インテリジェンスプロバイダーのテスト
+//!
+//! このモジュールには、各種脅威インテリジェンスプロバイダーの
+//! 単体テストと統合テストが含まれます。
+
+// プロバイダーテスト
+mod abuseipdb_provider_tests;
+mod cve_provider_tests;
+mod mitre_attack_provider_tests;
